requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
ddgs>=0.3.0